% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{check_duplicates}
\usage{
check_duplicates(dt = NULL, by_cols = NULL, noisily = T)
}
\arguments{
\item{dt}{A data.table.}

\item{by_cols}{A character vector of column names.}

\item{noisily}{Logical. If TRUE, print messages.}
}
\description{
Check if rows are uniquely identified by a set of columns.
}
