% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_merge.R
\name{check_merge}
\alias{check_merge}
\title{check_merge}
\arguments{
\item{data1}{is always the larger dataset and the one in which key combinations may repeat}

\item{data2}{is always uniquely identified by keys}

\item{keys}{is a vector of key columns}

\item{join_type}{is the type of join to perform: "outer"-default, "inner", "left", or "right"}

\item{abort}{is a logical that determines whether to abort if mismatches are detected}

\item{return_class}{is the class of the returned object: "data.table"-default, "tibble", or "data.frame"}

\item{keep_merge_vars}{is a logical that determines whether to keep the merge variables}

\item{noisily}{is a logical that determines whether to print output}
}
\description{
check_merge is a function that checks the results of a merge between two datasets
}
