% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miss.R
\name{check_miss}
\alias{check_miss}
\title{check_miss}
\usage{
check_miss(dt = NULL, exempt = NULL, noisily = T, abort = T)
}
\arguments{
\item{dt}{A data.table.}

\item{exempt}{A character vector of column names to be exempted from checks.}

\item{noisily}{Logical. If TRUE, print messages.}

\item{abort}{Logical. If TRUE, throw an error if missing values are found.}
}
\description{
Check for missing values in a data.table.
}
