% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddkcoidev-package.R
\docType{package}
\name{ddkcoidev-package}
\alias{ddkcoidev}
\alias{ddkcoidev-package}
\title{ddkcoidev}
\description{
ddkcoidev is an internal library to be used by our Data and Development team at Data Diversity Kids
for the development of the Child Opportunity Index (COI) and other related projects.
}
\details{
More detailed documentation here. We can also link to vignettes, describe main functions, describe
built in data, use examples, and provide references to other documentation.
}
\seealso{
git repo:  \url{https://github.com/briandevoe/ddkcoidev}
}
\author{
Brian DeVoe
}
\keyword{internal}
